Finally deleted swiftpp and qhull. Now we have less license problems. There is still license problem with math library, though.
Tetrahedrons are currently not working, just crashing (because swiftpp was removed).

SDECTriaxialTest is deleted, because it was never done, in fact.

+++ yade-guis/yade-gui-qt/src/QtGUI/QtFileGenerator.cpp
@@ -198,7 +198,7 @@ void QtFileGenerator::timerEvent( QTimerEvent* )
 	if(m_worker && m_runner)
 	{
 		textLabel1->setText(m_worker->message());
-		progressBar1->setProgress(m_worker->progress()*1000.0);
+		progressBar1->setProgress((int)(m_worker->progress()*1000.0));
 	}
 
 	// generation finished
